Backend/requirements.txt
fastapi==0.111.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.25.2
supabase==2.3.5
python-jose==3.3.0
python-multipart==0.0.7
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
jinja2==3.1.3
python-slugify==8.0.4 